Positional Targetting Complete + Mana Class
You can now target positions (I implemented the Move action as a test for this)
NOTE:: currently the ContTarget immediately executes actions once they are finished targetting.  This is nice for testing but should be removed.

The Mana class is pretty much done.  Not much that'll have to be extended from it.