Timeline Working
Basic timeline functionality is up and running.  Every 4 seconds it distributes a random mana, and executes any actions that are set for that turn.  (diagonal directions are move).
WARNING:  Since there's no gui, it's entirely possible there is slightly broken stuff that I can't detect.  Going to work on a View for this and the mana next.